fix(customizer-stack): reduce customizer stack to avoid issues
