+++ packages/agent/src/agent.ts
@@ -212,6 +212,7 @@ export default class Agent<S extends TSchema = TSchema> extends FrameworkMounter
     // It allows to rebuild the full customization stack with no code customizations
     this.nocodeCustomizer = new DataSourceCustomizer<S>({
       ignoreMissingSchemaElementErrors: this.options.ignoreMissingSchemaElementErrors || false,
+      strategy: 'NoCode',
     });
     this.nocodeCustomizer.addDataSource(this.customizer.getFactory());
     this.nocodeCustomizer.use(this.customizationService.addCustomizations);
